+++ tests/tests_fabric/test_fabric.py
@@ -718,15 +718,15 @@ def test_module_sharding_context():
     """Test that the sharding context manager gets applied when the strategy supports it and is a no-op
     otherwise."""
     fabric = Fabric()
-    fabric._strategy = MagicMock(spec=DDPStrategy, init_sharded_context=Mock())
-    with fabric.sharded_model():
+    fabric._strategy = MagicMock(spec=DDPStrategy, module_sharded_context=Mock())
+    with pytest.warns(DeprecationWarning, match="sharded_model"), fabric.sharded_model():
         pass
-    fabric._strategy.init_sharded_context.assert_not_called()
+    fabric._strategy.module_sharded_context.assert_not_called()
 
     fabric._strategy = MagicMock(spec=_Sharded)
-    with fabric.sharded_model():
+    with pytest.warns(DeprecationWarning, match="sharded_model"), fabric.sharded_model():
         pass
-    fabric._strategy.init_sharded_context.assert_called_once()
+    fabric._strategy.module_sharded_context.assert_called_once()
 
 
 def test_init_module_context(monkeypatch):
@@ -734,7 +734,7 @@ def test_init_module_context(monkeypatch):
     import lightning.fabric
 
     fabric = Fabric(accelerator="cpu")
-    strategy = lightning.fabric.strategies.SingleDeviceStrategy(device=torch.device("cuda"))
+    strategy = SingleDeviceStrategy(device=torch.device("cuda"))
     strategy.init_context = Mock(wraps=strategy.init_context)
     fabric._strategy = strategy
     with fabric.init_module():
@@ -743,11 +743,27 @@ def test_init_module_context(monkeypatch):
     strategy.init_context.reset_mock()
 
     # Pretend we are using PyTorch < 2.0
-    monkeypatch.setattr(lightning.fabric.strategies.strategy, "_TORCH_GREATER_EQUAL_2_0", False)
+    monkeypatch.setattr(lightning.fabric.fabric, "_TORCH_GREATER_EQUAL_2_0", False)
     with pytest.warns(PossibleUserWarning, match="can't place the model parameters on the device"):  # noqa: SIM117
         with fabric.init_module():
             pass
     strategy.init_context.assert_called_once()
+
+
+def test_init_context(monkeypatch):
+    """Test that `.init()` warns if using PyTorch < 2.0."""
+    # TODO(awaelchli): Extend the test once `Fabric.init()` finalized
+    import lightning.fabric
+
+    fabric = Fabric(accelerator="cpu")
+    strategy = SingleDeviceStrategy(device=torch.device("cuda"))
+    fabric._strategy = strategy
+
+    # Pretend we are using PyTorch < 2.0
+    monkeypatch.setattr(lightning.fabric.fabric, "_TORCH_GREATER_EQUAL_2_0", False)
+    with pytest.warns(PossibleUserWarning, match="can't place the model parameters on the device"):  # noqa: SIM117
+        with fabric.init():
+            pass
 
 
 def test_callbacks_input():